+++ src/HuffmanWeb/huffmanweb.client/src/Interfaces/Encode/BinaryHuffmanProps.tsx
