+++ src/HuffmanWeb/huffmanweb.client/src/Interfaces/Encode/CharactersListProps.tsx
