+++ src/HuffmanWeb/huffmanweb.client/src/Interfaces/Encode/JsonModalProps.tsx
