+++ src/HuffmanWeb/huffmanweb.client/src/Interfaces/Encode/SizeStatsProps.ts
