+++ src/HuffmanWeb/huffmanweb.client/src/Interfaces/Encode/TextToEncodeFormProps.tsx
