+++ src/HuffmanWeb/huffmanweb.client/src/Interfaces/Encode/TreeChildrenProps.tsx
