+++ src/HuffmanWeb/huffmanweb.client/src/Interfaces/Encode/TreeProps.tsx
