+++ src/HuffmanWeb/huffmanweb.client/src/Interfaces/MenuProps.tsx
@@ -1,0 +1,3 @@
+export interface MenuProps {
+  onSetMenu: (val: string) => void;
+}
