+++ src/HuffmanWeb/huffmanweb.client/src/dtos/Encode/TextToEncodeResponse.tsx
@@ -1,0 +1,10 @@
+import { Character } from "../Character";
+import { WeightedGraph } from "../WeightedGraph";
+
+export interface textToEncodeResponse {
+  graph: WeightedGraph;
+  matchingCharacters: Character[];
+  encodedSize: number;
+  originalSize: number;
+  encodedBinaryString: string;
+}
