+++ src/HuffmanWeb/huffmanweb.client/src/dtos/HuffmanNode.tsx
@@ -1,0 +1,5 @@
+export interface HuffmanNode {
+  identifier: string;
+  character: string;
+  nbOccurence: number;
+}
