+++ src/HuffmanWeb/huffmanweb.client/src/dtos/Link.tsx
@@ -1,0 +1,7 @@
+import { HuffmanNode } from "./HuffmanNode";
+
+export interface Link {
+  weight: number;
+  parent: HuffmanNode;
+  child: HuffmanNode;
+}
