+++ src/HuffmanWeb/huffmanweb.client/src/dtos/WeightedGraph.tsx
@@ -1,0 +1,8 @@
+import { HuffmanNode } from "./HuffmanNode";
+import { Link } from "./Link";
+
+export interface WeightedGraph {
+  root: HuffmanNode;
+  allNodes: HuffmanNode[];
+  links: Link[];
+}
